+++ crates/bevy_render/macros/src/specializer.rs
