+++ crates/bevy_render/src/render_resource/mod.rs
@@ -12,7 +12,7 @@ mod pipeline_cache;
 mod pipeline_specializer;
 pub mod resource_macros;
 mod shader;
-mod specialize;
+mod specializer;
 mod storage_buffer;
 mod texture;
 mod uniform_buffer;
@@ -29,7 +29,7 @@ pub use pipeline::*;
 pub use pipeline_cache::*;
 pub use pipeline_specializer::*;
 pub use shader::*;
-pub use specialize::*;
+pub use specializer::*;
 pub use storage_buffer::*;
 pub use texture::*;
 pub use uniform_buffer::*;
